fix: 增加部分注释以及修改部分代码位置
由于部分代码放置位置较乱,导致manager.go文件较大以及整体逻辑不清晰,因此进行优化.

Log: 增加部分注释以及修改部分代码位置
Task: https://pms.uniontech.com/task-view-216571.html
Influence: 注释和代码位置
