+++ src/lastore-daemon/update_infos.go
@@ -6,13 +6,7 @@ package main
 
 import (
 	"internal/system"
-	"os"
 	"path"
-	"time"
-
-	"github.com/linuxdeepin/go-lib/dbusutil"
-
-	"github.com/godbus/dbus"
 )
 
 type ApplicationInfo struct {
@@ -34,61 +28,6 @@ func (u *Updater) updateUpdatableApps() {
 	}
 	u.PropsMu.RUnlock()
 	u.setUpdatableApps(apps)
-}
-
-func (u *Updater) ApplicationUpdateInfos(lang string) (updateInfos []ApplicationUpdateInfo, busErr *dbus.Error) {
-	u.service.DelayAutoQuit()
-	iInfos := packageIconInfos()
-	aInfos := applicationInfos()
-	var uInfosMap system.SourceUpgradeInfoMap
-	var err error
-	repeatCount := 0
-	for {
-		if repeatCount > 5 {
-			break
-		}
-		uInfosMap, err = u.manager.SystemUpgradeInfo()
-		if os.IsNotExist(err) {
-			time.Sleep(1 * time.Second)
-			repeatCount++
-		} else if err != nil {
-			updateInfoErr, ok := err.(*system.UpdateInfoError)
-			if ok {
-				return nil, dbusutil.MakeErrorJSON(u, "UpdateInfoError", updateInfoErr)
-			}
-			return nil, dbusutil.ToError(err)
-		} else {
-			break
-		}
-	}
-
-	for _, uInfos := range uInfosMap {
-		for _, uInfo := range uInfos {
-			id := uInfo.Package
-
-			aInfo, ok := aInfos[id]
-			if !ok {
-				continue
-			}
-
-			info := ApplicationUpdateInfo{
-				Id:             id,
-				Name:           aInfo.LocaleName[lang],
-				Icon:           iInfos[id],
-				CurrentVersion: uInfo.CurrentVersion,
-				LastVersion:    uInfo.LastVersion,
-			}
-			if info.Name == "" {
-				info.Name = id
-			}
-			if info.Icon == "" {
-				info.Icon = id
-			}
-			updateInfos = append(updateInfos, info)
-		}
-	}
-	logger.Info("ApplicationUpdateInfos: ", updateInfos)
-	return updateInfos, nil
 }
 
 func applicationInfos() map[string]ApplicationInfo {
